Moved http/HeaderTools to HttpHeaderTools
to avoid a self-contradicting combination of
http/HeaderTools.cc and tests/stub_libhttp.cc
