+++ src/HttpHdrContRange.cc
@@ -12,7 +12,7 @@
 #include "base/Packable.h"
 #include "debug/Stream.h"
 #include "enums.h"
-#include "http/HeaderTools.h"
+#include "HttpHeaderTools.h"
 #include "HttpHdrContRange.h"
 
 /*
