+++ src/HttpHdrRange.cc
@@ -10,7 +10,7 @@
 
 #include "squid.h"
 #include "client_side_request.h"
-#include "http/HeaderTools.h"
+#include "HttpHeaderTools.h"
 #include "http/Stream.h"
 #include "HttpHeaderRange.h"
 #include "HttpReply.h"
