+++ src/client_side_reply.cc
@@ -22,7 +22,7 @@
 #include "FwdState.h"
 #include "globals.h"
 #include "HeaderMangling.h"
-#include "http/HeaderTools.h"
+#include "HttpHeaderTools.h"
 #include "http/Stream.h"
 #include "HttpReply.h"
 #include "HttpRequest.h"
