+++ src/http/Message.cc
@@ -11,7 +11,7 @@
 #include "squid.h"
 #include "debug/Stream.h"
 #include "http/ContentLengthInterpreter.h"
-#include "http/HeaderTools.h"
+#include "HttpHeaderTools.h"
 #include "http/Message.h"
 #include "http/one/Parser.h"
 #include "HttpHdrCc.h"
