+++ src/http/Stream.cc
@@ -9,7 +9,7 @@
 #include "squid.h"
 #include "client_side_request.h"
 #include "clientStream.h"
-#include "http/HeaderTools.h"
+#include "HttpHeaderTools.h"
 #include "http/Stream.h"
 #include "HttpHdrContRange.h"
 #include "Store.h"
